split Functions APIs into separate files in /_api/functions/
